Add support for Debian packages
This patch adds the required metadata for Debian packages to the
Cargo.toml.  (Unfortunately, it is AFAIS not possible to place this
metadata in the pkg subdirectory.)  It also adds a Makefile in
pkg/debian that invokes cargo deb to generate the Debian package.  The
package is placed in $CARGO_TARGET_DIR/debian.

The next step would be to set up a GitHub Action that builds the package
for every release and attaches it to the release on GitHub.